@@ -1,7 +1,10 @@
 package dto
 
-type UserResponse struct {
-	Status  int                    `json:"status"`
-	Message string                 `json:"message"`
-	Data    map[string]interface{} `json:"data"`
+type Response[T any] struct {
+	Status       int     `json:"status"`
+	Message      string  `json:"message"`
+	Data         T       `json:"data"`
+	ErrorMessage *string `json:"error"`
 }
+
+type UserResponse Response[interface{}]
